Moved SchedulerUpgrade startup routine to dhis-service-reporting
